Улучшена Архитектура приложения
