+++ README.md
@@ -1,5 +1,27 @@
-# Файлы для итогового задания
+Аналог TODO-листа, который реализует функциональность простейшего планировщика задач.
+
+Планировщик хранит задачи, каждая из них содержит дату дедлайна и заголовок с комментарием.
+Задачи могут повторяться по заданному правилу: например, ежегодно, через какое-то количество дней, в определённые дни месяца или недели.
+Если отметить такую задачу как выполненную, она переносится на следующую дату в соответствии с правилом. Обычные задачи при выполнении будут удаляться.
 
 В директории `tests` находятся тесты для проверки API, которое должно быть реализовано в веб-сервере.
 
 Директория `web` содержит файлы фронтенда.
+
+Все тесты проходят успешно с данными true в полях FullNextDate и Search в файле settings.go
+
+Выполенние задач со звездочкой:
+
+- Шаг 1. Запускаем веб-сервер Задание со `*` выполнено
+  (`Если существует переменная окружения TODO_PORT, сервер при старте слушает порт со значением этой переменной`)
+
+- Шаг 2. Проектируем и создаём БД Задание со `*` выполнено
+  (`Cервер получает значение переменной окружения TODO_DBFILE и использует его в качестве пути к базе данных (если это не пустая строка)`)
+
+- Шаг 3. Правила повторения задач Задание со `*` выполнено
+  (`Реализованы правила v и m`)
+  Шаг 5. Получаем список ближайших задач Задание со `*` выполнено
+  (`Добавлена возможность поиска задачи по дате и словам`)
+
+- Шаг 8. Заключительный. Задание со `*` `НЕ` выполнено
+  (`Данная тема была плохо раскрыта в обучении, и мне очень тяжело дается в понимании. Плюс после написания кода у меня перестали работать тесты, по этому обработчики без Аутентификации`)
