+++ internal/dates/nextdate.go
@@ -1,4 +1,4 @@
-package main
+package dates
 
 import (
 	"errors"
