+++ internal/db/db.go
@@ -1,4 +1,4 @@
-package main
+package db
 
 import (
 	"log"
