--- library.sql
@@ -1,7 +1,0 @@
-CREATE TABLE scheduler (
-    id INTEGER PRIMARY KEY AUTOINCREMENT,
-    date CHAR(8) NOT NULL DEFAULT '',
-    title VARCHAR(128) NOT NULL DEFAULT '',
-    comment VARCHAR(128) NOT NULL DEFAULT '',
-    repeat VARCHAR(128) NOT NULL DEFAULT ''
-);
